+++ src/Processors/BrowserProcessor.js
@@ -22,6 +22,12 @@ const BrowserProcessor = class extends ProcessorBase {
     }
   }
 
+  /** @inheritdoc */
+  getTrustedKeys() {
+    return ['browser'];
+  }
+
+
   /**
    * The only required method for processor implementations.
    *