+++ src/Senders/MongodbSender.js
@@ -1,6 +1,8 @@
 /**
  * @fileOverview MongoDB Sender class.
  */
+
+import * as util from "util";
 import SenderBase from "./SenderBase";
 
 /**
@@ -17,8 +19,10 @@ const MongodbSender = class extends SenderBase {
    *   The Meteor Mongo service.
    * @param {(String|Collection)} collection
    *   The collection or the name of the collection in which to log.
+   * @param {Object} formatOptions
+   *   Optional : The options used to format the message (default to { depth: 5 }).
    */
-  constructor(mongo, collection = "logger") {
+  constructor(mongo, collection = "logger", formatOptions = null) {
     super();
     if (collection instanceof mongo.Collection) {
       this.store = collection;
@@ -30,11 +34,16 @@ const MongodbSender = class extends SenderBase {
     else {
       throw new Error("MongodbSender requires a Collection or a collection name");
     }
+
+    this.formatOptions = formatOptions || {
+      depth: 5,
+      rawKeys: []
+    };
   }
 
   send(level, message, context) {
     let defaultedContext = context || {};
-    let doc = { level, message };
+    let doc = {level, message};
 
     // It should contain a timestamp object if it comes from ClientLogger.
     if (typeof defaultedContext.timestamp === "undefined") {
@@ -42,7 +51,13 @@ const MongodbSender = class extends SenderBase {
     }
     doc.context = defaultedContext;
     doc.context.timestamp.store = Date.now();
-    this.store.insert(doc);
+    try {
+      this.store.insert(doc);
+    }
+    catch (e) {
+      doc.context = util.inspect(doc, this.formatOptions);
+      this.store.insert(doc);
+    }
   }
 };