+++ src/ServerLogger.js
@@ -9,6 +9,8 @@ import * as util from "util";
  *
  * Its main method is log(level, message, context).
  *
+ * @extends Logger
+ *
  * @see ServerLogger.log
  */
 class ServerLogger extends Logger {
@@ -18,14 +20,16 @@ class ServerLogger extends Logger {
    *
    * @param {StrategyBase} strategy
    *   A logging strategy instance.
+   * @param {Processor[]} processors
+   *   An array of processor instances.
    * @param {Webapp} webapp
    *   The Meteor WebApp service.
    * @param {Object} parameters
    * - logRequestHeaders: add request headers to the log context. Defaults to true.
    * - servePath: the path on which to expose the logger endpoint. Defaults to "/logger".
    */
-  constructor(strategy, webapp = null, parameters = {}) {
-    super(strategy);
+  constructor(strategy, processors, webapp = null, parameters = {}) {
+    super(strategy, processors);
     const defaultParameters = {
       logRequestHeaders: true,
       servePath: "/logger"